requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

ffmpeg-python==0.2.0
fpdf==1.7.2
numpy==2.2.3
opencv-python==4.11.0.86
pytubefix==8.12.3
streamlit==1.42.2
streamlit-pdf-viewer==0.0.21
torch==2.6.0+cu126
torchvision==0.21.0+cu126
transformers==4.49.0
